+++ packages/ui/src/icons/creditCard.tsx
@@ -5,7 +5,8 @@ export function CreditCard(props: ComponentProps<"svg">) {
     <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 25 25" {...props}>
       <path
         d="M4.47974 16.618L16.4297 4.66797"
-        stroke="#A8ADB7"
+        stroke="currentColor"
+        fill="none"
         strokeWidth="1.5"
         strokeMiterlimit="10"
         strokeLinecap="round"
@@ -13,7 +14,8 @@ export function CreditCard(props: ComponentProps<"svg">) {
       />
       <path
         d="M11.6514 19.0179L12.8514 17.8179"
-        stroke="#A8ADB7"
+        stroke="currentColor"
+        fill="none"
         strokeWidth="1.5"
         strokeMiterlimit="10"
         strokeLinecap="round"
@@ -21,7 +23,8 @@ export function CreditCard(props: ComponentProps<"svg">) {
       />
       <path
         d="M14.343 16.3275L16.733 13.9375"
-        stroke="#A8ADB7"
+        stroke="currentColor"
+        fill="none"
         strokeWidth="1.5"
         strokeMiterlimit="10"
         strokeLinecap="round"
@@ -29,14 +32,16 @@ export function CreditCard(props: ComponentProps<"svg">) {
       />
       <path
         d="M4.15132 10.9778L10.7913 4.33777C12.9113 2.21777 13.9713 2.20777 16.0713 4.30777L20.9813 9.21777C23.0813 11.3178 23.0713 12.3778 20.9513 14.4978L14.3113 21.1378C12.1913 23.2578 11.1313 23.2678 9.03132 21.1678L4.12132 16.2578C2.02132 14.1578 2.02132 13.1078 4.15132 10.9778Z"
-        stroke="#A8ADB7"
+        stroke="currentColor"
+        fill="none"
         strokeWidth="1.5"
         strokeLinecap="round"
         strokeLinejoin="round"
       />
       <path
         d="M2.55005 22.7373H22.55"
-        stroke="#A8ADB7"
+        stroke="currentColor"
+        fill="none"
         strokeWidth="1.5"
         strokeLinecap="round"
         strokeLinejoin="round"