+++ packages/react-providers/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @polkadex/react-providers
 
+## 0.1.0
+
+### Minor Changes
+
+- da1a08f: Eslint dependency removed
+
+## 0.0.3
+
+### Patch Changes
+
+- d2f7f25: Trade account provider
+- Updated dependencies [d2f7f25]
+  - @polkadex/trade-wallet@1.0.0
+
 ## 0.0.2
 
 ### Patch Changes